requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.5
coverage==6.4.4
dj-database-url==1.0.0
Django==4.0.5
django-admin-interface==0.19.1
django-colorfield==0.7.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
gunicorn==20.1.0
iniconfig==1.1.1
joblib==1.1.0
mysqlclient==2.1.1
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
send-email==201212
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tomli==2.0.1
tzdata==2022.1
whitenoise==6.2.0
django-storages
boto3
